Dodano rzutowanie na int w Shape::lcolor

+++ Graph.h
@@ -157,7 +157,7 @@ public:
 	Shape& operator=(const Shape&) = delete;
 private:
 	vector<Point> points;	// not used by all shapes
-	Color lcolor {fl_color()};
+	Color lcolor {static_cast<int>(fl_color())};
 	Line_style ls {0};
 	Color fcolor {Color::invisible};
